feat: move right sidebar to general folder
