+++ src/components/General/RightSideBar/RightSidebar.stories.tsx
@@ -1,19 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import SearchInput from './SearchInput'
+import RightSidebar from './RightSidebar'
 
 const meta = {
-  title: 'Dashboard/SearchInput',
-  component: SearchInput,
+  title: 'General/RightSidebar',
+  component: RightSidebar,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof SearchInput>
+} satisfies Meta<typeof RightSidebar>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-  args: { value: 'Search value' },
+  args: {},
 }
