+++ src/components/General/RightSideBar/SearchInput.stories.tsx
@@ -1,19 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import ValidatorsList from './ValidatorsList'
+import SearchInput from './SearchInput'
 
 const meta = {
-  title: 'Dashboard/ValidatorsList',
-  component: ValidatorsList,
+  title: 'General/SearchInput',
+  component: SearchInput,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
-} satisfies Meta<typeof ValidatorsList>
+} satisfies Meta<typeof SearchInput>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-  args: {},
+  args: { value: 'Search value' },
 }
