+++ src/pages/Dashboard/Dashboard.tsx
@@ -2,7 +2,7 @@ import { XStack } from 'tamagui'
 
 import LeftSidebar from '../../components/General/LeftSidebar/LeftSidebar'
 import DashboardContent from './DashboardContent'
-import RightSidebar from './RightSideBar/RightSidebar'
+import RightSidebar from '../../components/General/RightSideBar/RightSidebar'
 
 const Dashboard = () => (
   <XStack style={{ height: '100vh' }}>
